requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
beautifulsoup4==4.13.3
blis==1.3.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
eval_type_backport==0.2.2
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
googlesearch-python==1.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.7
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
networkx==3.5
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.3.0
preshed==3.0.10
propcache==0.3.2
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
pymoex==0.0.4
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-telegram-bot==22.2
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rubicon-objc==0.4.9
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
selenium==4.23.1
sentence-transformers==5.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
tabulate==0.9.0
thinc==8.3.6
threadpoolctl==3.6.0
together==1.5.21
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
trio==0.26.0
trio-websocket==0.11.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
